refactor: CashLogService, WorkCompletionService, InterestService에서 도메인 서비스 로직 분리 (#211)
* refactor: WorkCompletionService에서 도메인 서비스 로직 분리

* refactor: InterestService에서 도메인 서비스 로직 분리

* refactor: CashLogService에서 도메인 서비스 로직 분리

* refactor: 사용 안하는 코드 정리, 클래스 패키지 이동
