+++ back/src/main/java/com/ll/goohaeyou/category/application/CategoryAdminService.java
@@ -11,7 +11,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
 public class CategoryAdminService {
     private final CategoryPolicy categoryPolicy;
     private final CategoryDomainService categoryDomainService;
