+++ back/src/main/java/com/ll/goohaeyou/category/application/CategoryService.java
@@ -6,11 +6,13 @@ import com.ll.goohaeyou.category.domain.service.CategoryDomainService;
 import com.ll.goohaeyou.category.domain.entity.Category;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class CategoryService {
     private final CategoryDomainService categoryDomainService;
