+++ back/src/main/java/com/ll/goohaeyou/jobPost/jobPost/application/InterestService.java
@@ -1,44 +1,35 @@
 package com.ll.goohaeyou.jobPost.jobPost.application;
 
-import com.ll.goohaeyou.global.exception.EntityNotFoundException;
-import com.ll.goohaeyou.jobPost.jobPost.domain.entity.Interest;
-import com.ll.goohaeyou.jobPost.jobPost.domain.entity.JobPostDetail;
-import com.ll.goohaeyou.jobPost.jobPost.domain.repository.JobPostDetailRepository;
-import com.ll.goohaeyou.member.member.domain.entity.Member;
 import com.ll.goohaeyou.global.event.notification.PostGetInterestedEvent;
-import com.ll.goohaeyou.auth.exception.AuthException;
-import com.ll.goohaeyou.member.member.domain.repository.MemberRepository;
+import com.ll.goohaeyou.jobPost.jobPost.domain.entity.JobPostDetail;
+import com.ll.goohaeyou.jobPost.jobPost.domain.policy.InterestPolicy;
+import com.ll.goohaeyou.jobPost.jobPost.domain.service.InterestDomainService;
+import com.ll.goohaeyou.jobPost.jobPost.domain.service.JobPostDetailDomainService;
+import com.ll.goohaeyou.member.member.domain.entity.Member;
+import com.ll.goohaeyou.member.member.domain.service.MemberDomainService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class InterestService {
-    private final MemberRepository memberRepository;
-    private final JobPostDetailRepository jobPostDetailRepository;
+    private final JobPostDetailDomainService jobPostDetailDomainService;
     private final ApplicationEventPublisher publisher;
+    private final MemberDomainService memberDomainService;
+    private final InterestPolicy interestPolicy;
+    private final InterestDomainService interestDomainService;
 
     @Transactional
     public void addInterestToPost(String username, Long postId) {
-        JobPostDetail postDetail = jobPostDetailRepository.findById(postId)
-                .orElseThrow(EntityNotFoundException.PostNotExistsException::new);
-        Member member = memberRepository.findByUsername(username)
-                .orElseThrow(EntityNotFoundException.MemberNotFoundException::new);
+        JobPostDetail postDetail = jobPostDetailDomainService.getById(postId);
+        Member member = memberDomainService.getByUsername(username);
 
-        if (hasInterest(postDetail, member)) {
-            throw new AuthException.NotAuthorizedException();
-        }
+        interestPolicy.validateCanInterest(postDetail, member);
 
-        postDetail.getInterests().add(
-                Interest.create(postDetail, member)
-        );
-
-        postDetail.getJobPost().increaseInterestCount();
+        interestDomainService.addInterest(postDetail, member);
 
         if (!postDetail.getAuthor().equals(username)) {
             publisher.publishEvent(new PostGetInterestedEvent(this, postDetail, member));
@@ -47,37 +38,17 @@ public class InterestService {
 
     @Transactional
     public void removeInterestFromPost(String username, Long postId) {
-        JobPostDetail postDetail = jobPostDetailRepository.findById(postId)
-                .orElseThrow(EntityNotFoundException.PostNotExistsException::new);
-        Member member = memberRepository.findByUsername(username)
-                .orElseThrow(EntityNotFoundException.MemberNotFoundException::new);
+        JobPostDetail postDetail = jobPostDetailDomainService.getById(postId);
+        Member member = memberDomainService.getByUsername(username);
 
-        if (!hasInterest(postDetail, member)) {
-            throw new AuthException.NotAuthorizedException();
-        }
+        interestPolicy.validateRemoveInterest(postDetail, member);
 
-        postDetail.getInterests().removeIf(interest -> interest.getMember().equals(member));
-        postDetail.getJobPost().decreaseInterestCount();
-    }
-
-    public boolean hasInterest(JobPostDetail post, Member member) {
-        return post.getInterests().stream()
-                .anyMatch(interest -> interest.getMember().equals(member));
+        interestDomainService.removeInterest(postDetail, member);
     }
 
     public boolean isInterested(String username, Long id) {
-        JobPostDetail postDetail = jobPostDetailRepository.findById(id)
-                .orElseThrow(EntityNotFoundException.PostNotExistsException::new);
-        List<String> interestedUsernames = getInterestedUsernames(postDetail);
+        JobPostDetail postDetail = jobPostDetailDomainService.getById(id);
 
-        return interestedUsernames.stream()
-                .anyMatch(username::equals);
-    }
-
-    public List<String> getInterestedUsernames(JobPostDetail jobPostDetail) {
-        return jobPostDetail.getInterests().stream()
-                .map(Interest::getMember)
-                .map(Member::getUsername)
-                .toList();
+        return interestDomainService.isInterested(username, postDetail);
     }
 }
